modulo login finalizado

+++ semear/src/main/java/com/projeto/semear/dao/EnderecoDao.java
@@ -3,5 +3,5 @@ package com.projeto.semear.dao;
 import com.projeto.semear.model.Endereco;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface EnderecoDao extends JpaRepository <Endereco, Long> {
+public interface EnderecoDao extends JpaRepository <Endereco, Integer> {
 }
